+++ lib/src/impl/common/summon.dart
@@ -71,6 +71,8 @@ abstract class SummonCommonClient implements SummonClient {
       event = new SummonAutoparkHeartbeatCarMessage(
           timestamp:
               new DateTime.fromMillisecondsSinceEpoch(message["timestamp"]));
+    } else if (msgType == "autopark:smart_summon_viz") {
+      event = new SummonVisualizationMessage(path: (message["path"] as List<dynamic>).cast<double>());
     } else {
       event = new SummonUnknownMessage(msgType, message);
     }